Added method of appending to CSV file instead of overwriting every time

+++ Statistical Program/gui.py
@@ -1,6 +1,6 @@
 from tkinter import *
-import csv
 import pandas as pd
+import os
 
 root = Tk()
 L1 = Label(root, text="Enter Data Separated by commas")
@@ -21,7 +21,10 @@ def toarr():
 
 def addtocsv():
     df = pd.DataFrame(tlist, columns=["colummn"])
-    df.to_csv('list.csv', index=False)
+    if not os.path.isfile('C:/GitRepo/Side-Projects/Statistical Program/list.csv'):
+        df.to_csv('list.csv', index=False)
+    else:
+        df.to_csv('list.csv', mode='a', header=False, index=False)
 
 
 button_disp = Button(root, text="Show", command=toarr)
